add initial support for k8s 1.18
k8s 1.18 had a lot of changes in the client-go library by including a
context.Context argument for the kube-apiserver interactions, for this
reason lots of files in cilium had to be modified to account for this
changes.

Some follow ups are still required since most of those context.Context
are now using context.TODO() which they require a deeper look
case-by-case.

Also, k8s.io/client-go/tools/metrics/metrics.go:L50-L59 has more metrics
that can be exposed which will also be done as a follow up, this commit
was entirely made to have an almost atomic upgrade of the k8s library
from 1.17 to 1.18 so we can start testing the compatibility with k8s
1.18.

The upstream tests were also failing because we have a delay to change
the pod identity when a label changes. This required a change in helm
chart to have this delay set to 0s in the upstream tests.

Signed-off-by: André Martins <andre@cilium.io>